🚚 Move print colour to display
